Moves contrllers sources to a specific folder
